fix tests

+++ src/test/java/br/com/tardelli/mapper/LogMapperTest.java
@@ -4,7 +4,6 @@ import br.com.tardelli.model.LogLap;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.math.BigDecimal;
 import java.time.LocalTime;
 
 public class LogMapperTest {
@@ -18,6 +17,6 @@ public class LogMapperTest {
     Assert.assertEquals(logLap.getLapTime(), LocalTime.parse("00:01:04.010"));
     Assert.assertEquals(logLap.getLapNumber(), new Integer(4));
     Assert.assertEquals(logLap.getHourLog(), LocalTime.parse("23:52:22.586"));
-    Assert.assertEquals(logLap.getAverageSpeed(), new BigDecimal("43.474"));
+    Assert.assertEquals(logLap.getAverageSpeed(), new Double("43.474"));
   }
 }
